requirements.txt
# This file contains the requirements needed to run the repository
python==3.7.7
tqdm==4.43.0
numpy==1.18.1
opencv-python==4.2.0.32
scikit-learn==0.23.1
tensorflow-cpu==2.2.0
scikit-metrics==0.1.0